First pass on the second iteration of the get area. Taking account of terrain for mouvement for example
Former-commit-id: 080ec8514cf8b5276f1da26abcbd99d114a214de

+++ GreenShells/GreenShells/PopUpWindow.cpp
@@ -42,7 +42,7 @@ PopUpWindow::PopUpWindow(std::string windowText, int width, int height)
 
 		lastPos = (pos == std::string::npos) ? pos : lastPos;
 		std::string aLine = windowText.substr(0, lastPos);
-		int numberOfSpceToFill = numberOfCharPerLine - aLine.length();
+		int numberOfSpceToFill = numberOfCharPerLine - static_cast<int>( aLine.length());
 		aLine.append(numberOfSpceToFill, ' ');
 		allLines.emplace_back(aLine);
 		windowText.erase(0, lastPos + delimiter.length());
